Distribution of multiple ego vehicles.
AgentManager on the master sends information about the current active agent to the clients.
ClientManager properly downloads all the agents' data when multiple of them are required.
NPCManager sends transform position and rotation instead of a rigidbody's position and rotation.
Fixed initial timescale in the SimulatorTimeManager.
Fixed adding distributed components to the agent's main rigidbody.
Agent selection is no longer interactable on the client simulation.