Improved DistributedObject identification.
Added IGloballyUniquelyIdentified interface.
DistributedObject uses IGloballyUniquelyIdentified's GUID as the key prefix instead of the hierarchy path if there is a component implementing the interface attached to the same gameobject.
NPCController, PedestrianController, IControllable and MapSignal implements IGloballyUniquelyIdentified .
PedestrianManager now sets the GTID of the PedestrianController when spawning the pedestrian.
DistributedRigidbody now uses coroutines instead of LateUpdate method.